+++ src/core/math/OverlapTester.hpp
@@ -29,7 +29,7 @@ public:
 
 private:
     static float calcAreaOfTriangle(float x1, float y1, float x2, float y2, float x3, float y3);
-    static bool doLinesIntersect(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4);
+    static bool doLinesIntersect(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4);
     
     OverlapTester();
     ~OverlapTester();
